Code update on google to latest revision as on 2nd Feb 2012. (revision 1598)

+++ trunk/dcma-ibm-cm/src/main/java/com/ephesoft/dcma/ibm/IBMCMExporter.java
@@ -1,37 +1,38 @@
-/********************************************************************************* 
-* Ephesoft is a Intelligent Document Capture and Mailroom Automation program 
-* developed by Ephesoft, Inc. Copyright (C) 2010-2011 Ephesoft Inc. 
-* 
-* This program is free software; you can redistribute it and/or modify it under 
-* the terms of the GNU Affero General Public License version 3 as published by the 
-* Free Software Foundation with the addition of the following permission added 
-* to Section 15 as permitted in Section 7(a): FOR ANY PART OF THE COVERED WORK 
-* IN WHICH THE COPYRIGHT IS OWNED BY EPHESOFT, EPHESOFT DISCLAIMS THE WARRANTY 
-* OF NON INFRINGEMENT OF THIRD PARTY RIGHTS. 
-* 
-* This program is distributed in the hope that it will be useful, but WITHOUT 
-* ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS 
-* FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more 
-* details. 
-* 
-* You should have received a copy of the GNU Affero General Public License along with 
-* this program; if not, see http://www.gnu.org/licenses or write to the Free 
-* Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 
-* 02110-1301 USA. 
-* 
-* You can contact Ephesoft, Inc. headquarters at 111 Academy Way, 
-* Irvine, CA 92617, USA. or at email address info@ephesoft.com. 
-* 
-* The interactive user interfaces in modified source and object code versions 
-* of this program must display Appropriate Legal Notices, as required under 
-* Section 5 of the GNU Affero General Public License version 3. 
-* 
-* In accordance with Section 7(b) of the GNU Affero General Public License version 3, 
-* these Appropriate Legal Notices must retain the display of the "Ephesoft" logo. 
-* If the display of the logo is not reasonably feasible for 
-* technical reasons, the Appropriate Legal Notices must display the words 
-* "Powered by Ephesoft". 
-********************************************************************************/ 
+/*********************************************************************************
+* Ephesoft is a Intelligent Document Capture and Mailroom Automation program
+* developed by Ephesoft, Inc. Copyright (C) 2010-2011 Ephesoft Inc.
+*
+* This program is free software; you can redistribute it and/or modify it under
+* the terms of the GNU Affero General Public License version 3 as published by the
+* Free Software Foundation with the addition of the following permission added
+* to Section 15 as permitted in Section 7(a): FOR ANY PART OF THE COVERED WORK
+* IN WHICH THE COPYRIGHT IS OWNED BY EPHESOFT, EPHESOFT DISCLAIMS THE WARRANTY
+* OF NON INFRINGEMENT OF THIRD PARTY RIGHTS.
+*
+* This program is distributed in the hope that it will be useful, but WITHOUT
+* ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+* FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more
+* details.
+*
+* You should have received a copy of the GNU Affero General Public License along with
+* this program; if not, see http://www.gnu.org/licenses or write to the Free
+* Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
+* 02110-1301 USA.
+*
+* You can contact Ephesoft, Inc. headquarters at 111 Academy Way,
+* Irvine, CA 92617, USA. or at email address info@ephesoft.com.
+*
+* The interactive user interfaces in modified source and object code versions
+* of this program must display Appropriate Legal Notices, as required under
+* Section 5 of the GNU Affero General Public License version 3.
+*
+* In accordance with Section 7(b) of the GNU Affero General Public License version 3,
+* these Appropriate Legal Notices must retain the display of the "Ephesoft" logo.
+* If the display of the logo is not reasonably feasible for
+* technical reasons, the Appropriate Legal Notices must display the words
+* "Powered by Ephesoft".
+********************************************************************************/
+
 
 package com.ephesoft.dcma.ibm;
 
@@ -86,7 +87,7 @@ import com.ephesoft.dcma.util.XMLUtil;
 
 /**
  * This Class is used to export the XML , DAT and CTL file in IBM content management accepted format.
- * 
+ *
  * @author Ephesoft
  * @version 1.0
  */
@@ -197,7 +198,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * Getter for supplyingSystem.
-	 * 
+	 *
 	 * @return {@link String}
 	 */
 
@@ -207,7 +208,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * Setter for supplyingSystem.
-	 * 
+	 *
 	 * @param supplyingSystem
 	 */
 	public final void setSupplyingSystem(final String supplyingSystem) {
@@ -244,7 +245,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This API update the batch.xml.
-	 * 
+	 *
 	 * @param batchInstanceID {@link String}
 	 */
 	private void updateBatch(final String batchInstanceID) {
@@ -266,7 +267,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This API retrieving the file location, calculating the file size and update it to batch.xml.
-	 * 
+	 *
 	 * @param batchInstanceID {@link String}
 	 * @param batch {@link Batch}
 	 * @param document {@link Document}
@@ -294,7 +295,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This API calculating the document size of mutipage pdf and returns the sum of all multipage pdf document.
-	 * 
+	 *
 	 * @param batchInstanceID {@link String}
 	 * @return totalDocSize
 	 */
@@ -320,7 +321,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This method transforms the batch.xml to another XML acceptable by IBM Content Management.
-	 * 
+	 *
 	 * @param batchInstanceID {@link String}
 	 * @throws JAXBException
 	 * @throws DCMAApplicationException
@@ -350,11 +351,16 @@ public class IBMCMExporter implements ICommonConstants {
 
 			LOGGER.info("Properties Initialized Successfully");
 
+			boolean isZipSwitchOn = batchSchemaService.isZipSwitchOn();
+			LOGGER.info("Zipped Batch XML switch is:" + isZipSwitchOn);
+
 			// check whether export folder to be created or not
 			String exportToFolder = batchSchemaService.getLocalFolderLocation();
 			String baseDocsFolder = exportToFolder + File.separator + batchInstanceID;
 			boolean isExportFolderCreated = folderCreation(exportFolder);
-			String sourceXMLPath = baseDocsFolder + File.separator + batchInstanceID + IBMCMConstant.BATCH_XML.getId();
+
+			String batchXmlName = batchInstanceID + ICommonConstants.UNDERSCORE_BATCH_XML;
+			String sourceXMLPath = baseDocsFolder + File.separator + batchInstanceID + ICommonConstants.UNDERSCORE_BATCH_XML;
 			Batch batch = batchSchemaService.getBatch(batchInstanceID);
 			String targetXmlPath = null;
 			String batchName = "";
@@ -376,16 +382,17 @@ public class IBMCMExporter implements ICommonConstants {
 				LOGGER.error("Could not find xsl file in the classpath resource", e.getMessage(), e);
 				throw new DCMAApplicationException("Could not find xsl file in the classpath resource", e);
 			}
-
+			InputStream in = null;
 			LOGGER.debug("Transforming XML " + sourceXMLPath + " to " + targetXmlPath);
 			if (xslStream != null && targetXmlPath != null && isExportFolderCreated) {
 				try {
 					TransformerFactory tFactory = TransformerFactory.newInstance();
 					Transformer transformer = null;
+
 					try {
 						transformer = tFactory.newTransformer(new StreamSource(xslStream));
 					} finally {
-						if(xslStream!= null) {
+						if (xslStream != null) {
 							try {
 								xslStream.close();
 							} catch (IOException e) {
@@ -435,8 +442,26 @@ public class IBMCMExporter implements ICommonConstants {
 						// Setting parameter for DAT file name for updating its value in XML.
 						transformer.setParameter(IBMCMConstant.DAT_FILE_NAME.getId(), new File(datFile).getName());
 						transformer.setParameter(IBMCMConstant.SUPLLYING_SYSTEM.getId(), this.getSupplyingSystem());
-						transformer.transform(new StreamSource(new File(sourceXMLPath)), new StreamResult(new FileOutputStream(
-								targetXmlPath)));
+
+						if (isZipSwitchOn) {
+							if (FileUtils.isZipFileExists(sourceXMLPath)) {
+								in = FileUtils.getInputStreamFromZip(sourceXMLPath, batchXmlName);
+								transformer.transform(new StreamSource(in), new StreamResult(new FileOutputStream(targetXmlPath)));
+							} else {
+								transformer.transform(new StreamSource(new File(sourceXMLPath)), new StreamResult(new FileOutputStream(
+										targetXmlPath)));
+							}
+						} else {
+							File srcXML = new File(sourceXMLPath);
+							if (srcXML.exists()) {
+								transformer.transform(new StreamSource(srcXML), new StreamResult(new FileOutputStream(
+										targetXmlPath)));
+							} else {
+								in = FileUtils.getInputStreamFromZip(sourceXMLPath, batchXmlName);
+								transformer.transform(new StreamSource(in), new StreamResult(new FileOutputStream(targetXmlPath)));
+							}
+						}
+
 						updateDocumentOffsetValue(batchInstanceID, targetXmlPath);
 					} else {
 						LOGGER.error("Transformer is null due to Invalid xsl file.");
@@ -447,7 +472,18 @@ public class IBMCMExporter implements ICommonConstants {
 				} catch (TransformerException e1) {
 					LOGGER.error("Problem occurred in transforming " + sourceXMLPath + " to " + targetXmlPath + e1.getMessage(), e1);
 					throw new DCMAApplicationException("Could not transform ibmCMTransform.xsl file : " + e1.getMessage(), e1);
-				} 
+				} catch (IOException ioe) {
+					LOGGER.error("Problem occurred in transforming " + sourceXMLPath + " to " + targetXmlPath + ioe.getMessage(), ioe);
+					throw new DCMAApplicationException("Could not transform ibmCMTransform.xsl file : " + ioe.getMessage(), ioe);
+				} finally {
+					if (in != null) {
+						try {
+							in.close();
+						} catch (IOException e) {
+							LOGGER.info("Error closing input stream for :" + sourceXMLPath);
+						}
+					}
+				}
 			} else {
 				LOGGER.error("Invalid input stream for :" + xslResource.toString());
 				throw new DCMAApplicationException("Could not find ibmCMTransform.xsl file");
@@ -464,7 +500,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This API is used to parsing parameter to ibmCMTransform.xsl file.
-	 * 
+	 *
 	 * @param batchInstanceID {@link String}
 	 * @param transformer {@link Transformer}
 	 */
@@ -509,7 +545,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This API created the export folder if it not exists else it will created the export folder.
-	 * 
+	 *
 	 * @param exportFolder {@link String}
 	 * @return
 	 */
@@ -527,7 +563,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This API provides the target file path to be created.
-	 * 
+	 *
 	 * @param batchInstanceID {@link String}
 	 * @param exportFolder {@link String}
 	 * @param batch {@link Batch}
@@ -565,7 +601,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This API fetching the document list of mutipage pdf.
-	 * 
+	 *
 	 * @param batchInstanceID {@link String}
 	 * @return totalDocSize
 	 */
@@ -593,7 +629,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This API for generating DAT file.
-	 * 
+	 *
 	 * @param batchInstanceID {@link String}
 	 * @param exportFolder {@link String}
 	 * @param batch {@link Batch}
@@ -649,7 +685,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This API for generating CTL file.
-	 * 
+	 *
 	 * @param batchInstanceID {@link String}
 	 * @param exportFolder {@link String}
 	 * @param batch {@link Batch}
@@ -680,7 +716,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This API is used to update the offset value in XML file generated for IBM Content Management.
-	 * 
+	 *
 	 * @param batchInstanceID {@link String}
 	 * @param filePath {@link String}
 	 * @throws DCMAApplicationException if any exception or error occur
@@ -730,7 +766,7 @@ public class IBMCMExporter implements ICommonConstants {
 
 	/**
 	 * This API is used to get the offset value for the specific document identifier.
-	 * 
+	 *
 	 * @param batchInstanceID {@link String}
 	 * @param docID {@link String}
 	 * @return offset
@@ -756,6 +792,11 @@ public class IBMCMExporter implements ICommonConstants {
 		return offset;
 	}
 
+	/**
+	 * API for getting subpoena loan number on the basis of document list.
+	 * @param documentList
+	 * @return
+	 */
 	private Map<String, String> getSubPoenaLoanNumber(List<Document> documentList) {
 		Map<String, String> resultMap = new HashMap<String, String>();
 		for (Document document : documentList) {
